move tipper into its own component
